fix example 012

+++ examples/012-can/f107vc/src/main.c
@@ -3,8 +3,8 @@
 // =============================================================================
 static void taskCan1Test(void) {
     uint8_t data[8];
-    for (int i = 0; i < 8; i++) {
-        data[i] = (i + 1) * 16 + (i + 1);
+    for (uint8_t i = 0; i < 8; i++) {
+        data[i] = ( uint8_t )((i + 1) * 16 + (i + 1));
     }
 
     uint16_t id = 0x555;
